troubleshooting issues with the controller.
